requirements.txt
boto3>=1.34.0
pandas>=2.0.0
click>=8.1.0
python-dotenv>=1.0.0
tqdm>=4.66.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
check-orientation>=0.0.5